backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
aiosqlite==0.19.0
greenlet==3.3.2
httpx==0.25.2
openai==1.6.1
websockets==12.0
python-multipart==0.0.6
